Moving chat message logic into separate classes
