+++ app/Service/Chat/ChatService.php
@@ -84,129 +84,6 @@ class ChatService
     }
 
     /**
-     * Send message service
-     *
-     * @param array $data
-     * @return array
-     */
-    public function send(array $data): array
-    {
-        $dialogWithId = (int)$data['dialogWithId'];
-        $dialogId = (int)$data['dialogId'];
-        $message = trim($data['message']);
-
-        $this->checkDialogAccess($dialogId);
-
-        $messageId = $this->chatRepository->sendMessage(
-            $message,
-            $this->getDialogId($dialogWithId, $dialogId)
-        );
-
-        if (!$messageId) {
-            throw new ChatMessageException('Message was not sent!');
-        }
-
-        // Websocket
-        $message = MessagesModel::where('message_id', $messageId)->firstOrFail();
-        $user = User::select(['id', 'name', 'avatar'])
-            ->whereId(auth()->id())
-        ->firstOrFail();
-        broadcast(new \App\Events\ChatMessageEvent($user, $message));
-
-        return [
-            'id' => $messageId,
-            'created_at' => $message->created_at
-        ];
-    }
-
-    /**
-     * Edit message in dialog
-     *
-     * @param array $data
-     * @return array
-     */
-    public function edit(array $data):array
-    {
-        $dialogId = (int) $data['dialogId'];
-        $messageId = (int) $data['messageId'];
-        $message = trim($data['message']);
-
-        $this->checkDialogAccess($dialogId);
-
-        $messageObj = MessagesModel::where('message_id', $messageId)
-            ->where('dialog_id', $dialogId)
-            ->firstOrFail();
-        $messageObj->text = $message;
-
-        if (!$messageObj->save()) {
-            throw new ChatMessageException('The message has not been changed!');
-        }
-
-        return [
-            'id' => $messageId,
-            'updated_at' => $messageObj->updated_at
-        ];
-    }
-
-    /**
-     * Delete message in dialog
-     *
-     * @param array $data
-     * @return array
-     */
-    public function delete(array $data):array
-    {
-        $dialogId = (int) $data['dialogId'];
-        $messageId = (int) $data['messageId'];
-
-        $this->checkDialogAccess($dialogId);
-
-        $messageObj = MessagesModel::query()
-            ->where('message_id', $messageId)
-            ->where('dialog_id', $dialogId)
-            ->firstOrFail();
-        $messageObj->delete();
-
-        if (!$messageObj->save()) {
-            throw new ChatMessageException('The message was not deleted!');
-        }
-
-        return [
-            'id' => $messageId,
-            'deleted_at' => $messageObj->deleted_at
-        ];
-    }
-
-    /**
-     * Recover message in dialog
-     *
-     * @param array $data
-     * @return
-     */
-    public function recover(array $data):array
-    {
-        $dialogId = (int) $data['dialogId'];
-        $messageId = (int) $data['messageId'];
-
-        $this->checkDialogAccess($dialogId);
-
-        $messageObj = MessagesModel::onlyTrashed()
-            ->where('message_id', $messageId)
-            ->where('dialog_id', $dialogId)
-            ->firstOrFail();
-        $messageObj->restore();
-
-        if (!$messageObj->save()) {
-            throw new ChatMessageException('The message was not restored!');
-        }
-
-        return [
-            'id' => $messageId,
-            'updated_at' => $messageObj->updated_at
-        ];
-    }
-
-    /**
      * User dialog service
      *
      * @param int $dialogId
@@ -296,7 +173,7 @@ class ChatService
     }
 
     /**
-     * Format create date message value
+     * Formatting the text of the message composition
      *
      * @param \stdClass $obj
      *
